Enhance README and implement ColorChangerScreen widget for dynamic background color change
